+++ frontend/src/Book/Pool/BookPoolAuthorIndexConnector.js
@@ -1,0 +1,150 @@
+/* eslint max-params: 0 */
+import PropTypes from 'prop-types';
+import React, { Component } from 'react';
+import { connect } from 'react-redux';
+import { createSelector } from 'reselect';
+import * as commandNames from 'Commands/commandNames';
+import withScrollPosition from 'Components/withScrollPosition';
+import { mergeAuthors, saveAuthorEditor, setAuthorFilter, setAuthorSort, setAuthorTableOption, setAuthorView } from 'Store/Actions/authorIndexActions';
+import { executeCommand } from 'Store/Actions/commandActions';
+import { fetchBookPoolAuthors } from 'Store/Actions/bookPoolAuthorActions';
+import scrollPositions from 'Store/scrollPositions';
+import createBookPoolAuthorClientSideCollectionItemsSelector from 'Store/Selectors/createBookPoolAuthorClientSideCollectionItemsSelector';
+import createCommandExecutingSelector from 'Store/Selectors/createCommandExecutingSelector';
+import createDimensionsSelector from 'Store/Selectors/createDimensionsSelector';
+import AuthorIndex from 'Author/Index/AuthorIndex';
+import BookPoolAuthorIndexFooterConnector from './BookPoolAuthorIndexFooterConnector';
+
+function createMapStateToProps() {
+  return createSelector(
+    createBookPoolAuthorClientSideCollectionItemsSelector('authorIndex'),
+    createCommandExecutingSelector(commandNames.BULK_REFRESH_AUTHOR),
+    createCommandExecutingSelector(commandNames.RENAME_AUTHOR),
+    createCommandExecutingSelector(commandNames.RETAG_AUTHOR),
+    createDimensionsSelector(),
+    (
+      author,
+      isRefreshingAuthor,
+      isOrganizingAuthor,
+      isRetaggingAuthor,
+      dimensionsState
+    ) => {
+      return {
+        ...author,
+        isRefreshingAuthor,
+        isOrganizingAuthor,
+        isRetaggingAuthor,
+        isSmallScreen: dimensionsState.isSmallScreen
+      };
+    }
+  );
+}
+
+function createMapDispatchToProps(dispatch, props) {
+  return {
+    dispatchFetchBookPoolAuthors() {
+      dispatch(fetchBookPoolAuthors());
+    },
+
+    onTableOptionChange(payload) {
+      dispatch(setAuthorTableOption(payload));
+    },
+
+    onSortSelect(sortKey) {
+      dispatch(setAuthorSort({ sortKey }));
+    },
+
+    onFilterSelect(selectedFilterKey) {
+      dispatch(setAuthorFilter({ selectedFilterKey }));
+    },
+
+    dispatchSetAuthorView(view) {
+      dispatch(setAuthorView({ view }));
+    },
+
+    dispatchSaveAuthorEditor(payload) {
+      dispatch(saveAuthorEditor(payload));
+    },
+
+    dispatchMergeAuthors(payload) {
+      dispatch(mergeAuthors(payload));
+    },
+
+    onRefreshAuthorPress(items) {
+      dispatch(executeCommand({
+        name: commandNames.BULK_REFRESH_AUTHOR,
+        authorIds: items,
+        skipNewBooks: true
+      }));
+    }
+  };
+}
+
+class BookPoolAuthorIndexConnector extends Component {
+
+  componentDidMount() {
+    this.props.dispatchFetchBookPoolAuthors();
+    if (typeof window !== 'undefined') {
+      window.addEventListener('bookPoolResourceUpdated', this.onBookPoolResourceUpdated);
+    }
+  }
+
+  componentWillUnmount() {
+    if (typeof window !== 'undefined') {
+      window.removeEventListener('bookPoolResourceUpdated', this.onBookPoolResourceUpdated);
+    }
+  }
+
+  //
+  // Listeners
+
+  onViewSelect = (view) => {
+    this.props.dispatchSetAuthorView(view);
+  };
+
+  onSaveSelected = (payload) => {
+    this.props.dispatchSaveAuthorEditor(payload);
+  };
+
+  onMergeAuthors = (payload) => {
+    this.props.dispatchMergeAuthors(payload);
+  };
+
+  onBookPoolResourceUpdated = () => {
+    this.props.dispatchFetchBookPoolAuthors();
+  };
+
+  onScroll = ({ scrollTop }) => {
+    scrollPositions.bookPoolAuthorIndex = scrollTop;
+  };
+
+  //
+  // Render
+
+  render() {
+    return (
+      <AuthorIndex
+        {...this.props}
+        footerComponent={BookPoolAuthorIndexFooterConnector}
+        onViewSelect={this.onViewSelect}
+        onScroll={this.onScroll}
+        onSaveSelected={this.onSaveSelected}
+        onMergeAuthors={this.onMergeAuthors}
+      />
+    );
+  }
+}
+
+BookPoolAuthorIndexConnector.propTypes = {
+  dispatchFetchBookPoolAuthors: PropTypes.func.isRequired,
+  isSmallScreen: PropTypes.bool.isRequired,
+  view: PropTypes.string.isRequired,
+  dispatchSetAuthorView: PropTypes.func.isRequired,
+  dispatchSaveAuthorEditor: PropTypes.func.isRequired,
+  dispatchMergeAuthors: PropTypes.func.isRequired
+};
+
+export default withScrollPosition(
+  connect(createMapStateToProps, createMapDispatchToProps)(BookPoolAuthorIndexConnector),
+  'bookPoolAuthorIndex'
+);