+++ frontend/src/Store/scrollPositions.js
@@ -1,5 +1,6 @@
 const scrollPositions = {
   authorIndex: 0,
+  bookPoolAuthorIndex: 0,
   bookIndex: 0
 };